Unit tests working.

+++ LikeMagic/Source/binding_functions.cpp
@@ -29,7 +29,7 @@ LM::TypeMirror& register_namespace(std::string name, TypeMirror& parent_namespac
     }
     else
     {
-        auto result = new LM::TypeMirror(full_name, 0, namespace_type);
+        auto result = new LM::TypeMirror(full_name, size_t(0), namespace_type);
         type_system->add_class(namespace_type, result, parent_namespace_);
         return *result;
     }
